+++ samples/Sentry.Samples.AspNetCore.Mvc/Controllers/HomeController.cs
@@ -1,6 +1,7 @@
 using System.Data;
 using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
+using Sentry.Ben.BlockingDetector;
 using Sentry.Samples.AspNetCore.Mvc.Models;
 
 namespace Samples.AspNetCore.Mvc.Controllers;
@@ -10,6 +11,53 @@ public class HomeController(ILogger<HomeController> logger) : Controller
     public IActionResult Index()
     {
         return View();
+    }
+
+    // GET /home/block/true or /home/block/false to observe events
+    [HttpGet("[controller]/block/{block?}")]
+    public async Task<string> Block([FromRoute] bool block)
+    {
+        if (block)
+        {
+            logger.LogInformation("\ud83d\ude31 Calling a blocking API on an async method \ud83d\ude31");
+
+            // This will result in an event in Sentry
+            Task.Delay(10).Wait(); // This is a blocking call. Same with '.Result'
+        }
+        else
+        {
+            logger.LogInformation("\ud83d\ude31 No blocking call made \ud83d\ude31");
+
+            // Non-blocking, no event captured
+            await Task.Delay(10);
+        }
+
+        return "Was blocking? " + block;
+    }
+
+    // GET /home/suppress/true or /home/suppress/false to observe events
+    [HttpGet("[controller]/suppress/{suppress?}")]
+    public async Task<string> Suppress([FromRoute] bool suppress)
+    {
+        if (suppress)
+        {
+            logger.LogInformation("Blocking suppression enabled");
+            using (new SuppressBlockingDetection())
+            {
+                Task.Delay(10).Wait(); // This is blocking but won't trigger an event, due to suppression
+            }
+            logger.LogInformation("Blocking suppression disabled");
+        }
+        else
+        {
+            logger.LogInformation("\ud83d\ude31 Unsuppressed blocking call on an async method \ud83d\ude31");
+            Task.Delay(10).Wait(); // This is blocking but won't trigger an event, due to suppression
+        }
+
+        // Non-blocking, no event captured
+        await Task.Delay(10);
+
+        return "Was suppressed? " + suppress;
     }
 
     // Example: An exception that goes unhandled by the app will be captured by Sentry: